Moved dummy employee records endpoint in app, slight refactor
